# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Email Operations Workspace API
  version: 1.0.0
servers:
  - url: /api/v1
security:
  - sessionCookie: []
paths:
...
  /recipient-lists:
    get:
      operationId: listRecipientLists
      description: 'BR-SEG-001/004/007/010: tenant-scoped reusable recipient lists with server-computed active-recipient counts.'
      parameters:
        - {$ref: '#/components/parameters/SegmentSearch'}
        - {$ref: '#/components/parameters/SegmentCursor'}
        - {$ref: '#/components/parameters/SegmentLimit'}
      responses:
        '200':
          description: Stable name/id ordered page of recipient lists
          content:
            application/json:
              schema: {$ref: '#/components/schemas/RecipientListCatalogResponse'}
        '403': {$ref: '#/components/responses/Problem'}
    post:
      operationId: createRecipientList
      description: 'BR-SEG-002: name is trimmed and case-insensitively unique among active lists in the current tenant.'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/RecipientListCreateRequest'}
      responses:
        '201':
          description: Recipient list created
          content:
            application/json:
              schema: {$ref: '#/components/schemas/RecipientList'}
        '400': {$ref: '#/components/responses/Problem'}
        '403': {$ref: '#/components/responses/Problem'}
        '409': {$ref: '#/components/responses/Problem'}
  /recipient-lists/{recipientListId}:
    get:
      operationId: getRecipientList
      parameters: [{$ref: '#/components/parameters/RecipientListId'}]
      responses:
        '200':
          description: Recipient list
          content:
            application/json:
              schema: {$ref: '#/components/schemas/RecipientList'}
        '403': {$ref: '#/components/responses/Problem'}
        '404': {$ref: '#/components/responses/Problem'}
    patch:
      operationId: updateRecipientList
      parameters: [{$ref: '#/components/parameters/RecipientListId'}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/RecipientListUpdateRequest'}
      responses:
        '200':
          description: Recipient list updated
          content:
            application/json:
              schema: {$ref: '#/components/schemas/RecipientList'}
        '400': {$ref: '#/components/responses/Problem'}
        '403': {$ref: '#/components/responses/Problem'}
        '404': {$ref: '#/components/responses/Problem'}
        '409': {$ref: '#/components/responses/Problem'}
    delete:
      operationId: deleteRecipientList
      description: 'BR-SEG-005: soft-deletes a list and removes only its memberships. Refuses a list referenced by a scheduled campaign snapshot.'
      parameters: [{$ref: '#/components/parameters/RecipientListId'}]
      responses:
        '204': {description: List soft-deleted; recipients and campaign snapshots are unchanged}
        '403': {$ref: '#/components/responses/Problem'}
        '404': {$ref: '#/components/responses/Problem'}
        '409': {$ref: '#/components/responses/Problem'}
  /recipient-lists/{recipientListId}/members:
    post:
      operationId: addRecipientListMembers
      parameters: [{$ref: '#/components/parameters/RecipientListId'}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/SegmentMembersRequest'}
      responses: {'201': {description: Memberships added idempotently}, '400': {$ref: '#/components/responses/Problem'}, '403': {$ref: '#/components/responses/Problem'}, '404': {$ref: '#/components/responses/Problem'}}
    delete:
      operationId: removeRecipientListMembers
      parameters: [{$ref: '#/components/parameters/RecipientListId'}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/SegmentMembersRequest'}
      responses: {'204': {description: Memberships removed}, '400': {$ref: '#/components/responses/Problem'}, '403': {$ref: '#/components/responses/Problem'}, '404': {$ref: '#/components/responses/Problem'}}
  /recipients/{recipientId}/segments:
    get:
      operationId: getRecipientSegments
      description: 'BR-SEG-007 / BR-REC-005/006: canonical list/tag membership for a recipient, consumed by the recipient table and form.'
      parameters: [{$ref: '#/components/parameters/RecipientId'}]
      responses:
        '200':
          description: Recipient segment memberships
          content:
            application/json:
              schema: {$ref: '#/components/schemas/RecipientSegments'}
        '403': {$ref: '#/components/responses/Problem'}
        '404': {$ref: '#/components/responses/Problem'}
  /tags:
    get:
      operationId: listTags
      description: 'BR-SEG-001/003/004/007/010: tenant-scoped flexible tags with server-computed active-recipient counts.'
      parameters:
        - {$ref: '#/components/parameters/SegmentSearch'}
        - {$ref: '#/components/parameters/SegmentCursor'}
        - {$ref: '#/components/parameters/SegmentLimit'}
      responses:
        '200':
          description: Stable name/id ordered page of tags
          content:
            application/json:
              schema: {$ref: '#/components/schemas/TagCatalogResponse'}
        '403': {$ref: '#/components/responses/Problem'}
    post:
      operationId: createTag
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/TagCreateRequest'}
      responses:
        '201':
          description: Tag created
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Tag'}
        '400': {$ref: '#/components/responses/Problem'}
        '403': {$ref: '#/components/responses/Problem'}
        '409': {$ref: '#/components/responses/Problem'}
  /tags/{tagId}:
    get:
      operationId: getTag
      parameters: [{$ref: '#/components/parameters/TagId'}]
      responses:
        '200':
          description: Tag
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Tag'}
        '403': {$ref: '#/components/responses/Problem'}
        '404': {$ref: '#/components/responses/Problem'}
    patch:
      operationId: updateTag
      parameters: [{$ref: '#/components/parameters/TagId'}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/TagUpdateRequest'}
      responses:
        '200':
          description: Tag updated
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Tag'}
        '400': {$ref: '#/components/responses/Problem'}
        '403': {$ref: '#/components/responses/Problem'}
        '404': {$ref: '#/components/responses/Problem'}
        '409': {$ref: '#/components/responses/Problem'}
    delete:
      operationId: deleteTag
      description: 'BR-SEG-006: soft-deletes the tag and removes only tag memberships; recipients and custom data remain unchanged.'
      parameters: [{$ref: '#/components/parameters/TagId'}]
      responses: {'204': {description: Tag soft-deleted}, '403': {$ref: '#/components/responses/Problem'}, '404': {$ref: '#/components/responses/Problem'}}
  /tags/{tagId}/members:
    post:
      operationId: addTagMembers
      parameters: [{$ref: '#/components/parameters/TagId'}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/SegmentMembersRequest'}
      responses: {'201': {description: Memberships added idempotently}, '400': {$ref: '#/components/responses/Problem'}, '403': {$ref: '#/components/responses/Problem'}, '404': {$ref: '#/components/responses/Problem'}}
    delete:
      operationId: removeTagMembers
      parameters: [{$ref: '#/components/parameters/TagId'}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/SegmentMembersRequest'}
      responses: {'204': {description: Memberships removed}, '400': {$ref: '#/components/responses/Problem'}, '403': {$ref: '#/components/responses/Problem'}, '404': {$ref: '#/components/responses/Problem'}}
...
  /bulk-jobs:
    get:
      operationId: listBulkJobs
      description: "Lists the current tenant's recent bulk-update jobs. This canonical history backs the recipient result UI; realtime only prompts a refetch."
      responses:
        '200':
          description: Tenant-scoped bulk job history
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/BulkJob'}
        '401': {$ref: '#/components/responses/Problem'}
        '403': {$ref: '#/components/responses/Problem'}
    post:
      operationId: createBulkJob
      description: 'M2-S4 / BR-GEN-005, BR-CF-004/005/006/007, BR-REC-009: resolves and freezes an explicit recipient selection before queuing the bulk operation.'
      parameters:
        - {$ref: '#/components/parameters/IdempotencyKey'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/BulkJobCreateRequest'}
      responses:
        '202':
          description: Bulk job accepted
          content:
            application/json:
              schema: {$ref: '#/components/schemas/BulkJob'}
        '400': {$ref: '#/components/responses/Problem'}
        '403': {$ref: '#/components/responses/Problem'}
        '409': {$ref: '#/components/responses/Problem'}
  /bulk-jobs/preview:
    post:
      operationId: previewBulkJob
      description: 'BR-CF-004 authoritative dry-run: validates the action and returns the resolved selected-recipient scope without creating a job or checkpoint.'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/BulkJobCreateRequest'}
      responses:
        '200':
          description: Validated bulk scope estimate
          content:
            application/json:
              schema: {$ref: '#/components/schemas/BulkJobPreview'}
        '400': {$ref: '#/components/responses/Problem'}
        '403': {$ref: '#/components/responses/Problem'}
        '404': {$ref: '#/components/responses/Problem'}
  /bulk-jobs/{bulkJobId}:
    get:
      operationId: getBulkJob
      parameters:
        - {$ref: '#/components/parameters/BulkJobId'}
      responses:
        '200':
          description: Bulk job
          content:
            application/json:
              schema: {$ref: '#/components/schemas/BulkJob'}
        '403': {$ref: '#/components/responses/Problem'}
        '404': {$ref: '#/components/responses/Problem'}
  /bulk-jobs/{bulkJobId}/error-file:
    get:
      operationId: downloadBulkErrorFile
      description: 'Downloads failed bulk-update recipient checkpoints with a machine-readable error reason. The job remains tenant-scoped.'
      parameters:
        - {$ref: '#/components/parameters/BulkJobId'}
      responses:
        '200':
          description: CSV file of failed recipient checkpoints; header-only when the job has no failures.
          content:
            text/csv:
              schema: {type: string, format: binary}
        '403': {$ref: '#/components/responses/Problem'}
        '404': {$ref: '#/components/responses/Problem'}
  /bulk-jobs/{bulkJobId}/result-file:
    get:
      operationId: downloadBulkResultFile
      description: 'Downloads the tenant-scoped CSV result for a completed bulk export. Other bulk job types do not expose this artifact.'
      parameters:
        - {$ref: '#/components/parameters/BulkJobId'}
      responses:
        '200':
          description: CSV file containing the successful frozen recipient checkpoints for the export.
          content:
            text/csv:
              schema: {type: string, format: binary}
        '403': {$ref: '#/components/responses/Problem'}
        '404': {$ref: '#/components/responses/Problem'}
...
components:
  securitySchemes:
    sessionCookie: {type: apiKey, in: cookie, name: eow_session}
  parameters:
...
    RecipientId: {name: recipientId, in: path, required: true, schema: {type: string, format: uuid}}
    RecipientListId: {name: recipientListId, in: path, required: true, schema: {type: string, format: uuid}}
    TagId: {name: tagId, in: path, required: true, schema: {type: string, format: uuid}}
...
    BulkJobId: {name: bulkJobId, in: path, required: true, schema: {type: string, format: uuid}}
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: {type: string, minLength: 1, maxLength: 255}
      description: 'Required by BR-GEN-005 so every import/bulk job creation is retry-safe.'
    SegmentSearch: {name: search, in: query, required: false, schema: {type: string, maxLength: 200}}
    SegmentCursor: {name: cursor, in: query, required: false, schema: {type: string}}
    SegmentLimit: {name: limit, in: query, required: false, schema: {type: integer, minimum: 1, maximum: 100, default: 25}}
...
  schemas:
...
    RecipientList:
      type: object
      required: [id, name, description, memberCount, createdAt, updatedAt]
      properties:
        id: {type: string, format: uuid}
        name: {type: string}
        description: {type: [string, 'null']}
        memberCount: {type: integer, minimum: 0}
        createdAt: {type: string, format: date-time}
        updatedAt: {type: string, format: date-time}
    RecipientListCatalogResponse:
      type: object
      required: [items, nextCursor, total]
      properties:
        items: {type: array, items: {$ref: '#/components/schemas/RecipientList'}}
        nextCursor: {type: [string, 'null']}
        total: {type: integer, minimum: 0}
    RecipientListCreateRequest:
      type: object
      required: [name]
      properties:
        name: {type: string, minLength: 1, maxLength: 160}
        description: {type: [string, 'null'], maxLength: 1000}
    RecipientListUpdateRequest:
      type: object
      properties:
        name: {type: string, minLength: 1, maxLength: 160}
        description: {type: [string, 'null'], maxLength: 1000}
    Tag:
      type: object
      required: [id, name, color, memberCount, createdAt, updatedAt]
      properties:
        id: {type: string, format: uuid}
        name: {type: string}
        color: {$ref: '#/components/schemas/TagColor'}
        memberCount: {type: integer, minimum: 0}
        createdAt: {type: string, format: date-time}
        updatedAt: {type: string, format: date-time}
    TagCatalogResponse:
      type: object
      required: [items, nextCursor, total]
      properties:
        items: {type: array, items: {$ref: '#/components/schemas/Tag'}}
        nextCursor: {type: [string, 'null']}
        total: {type: integer, minimum: 0}
    TagColor:
      type: string
      enum: ['#ef6f45', '#7356c8', '#278b6e', '#d79022', '#3a78c2', '#9a5eb0', '#31806b', '#b85b73']
    TagCreateRequest:
      type: object
      required: [name, color]
      properties:
        name: {type: string, minLength: 1, maxLength: 160}
        color: {$ref: '#/components/schemas/TagColor'}
    TagUpdateRequest:
      type: object
      properties:
        name: {type: string, minLength: 1, maxLength: 160}
        color: {$ref: '#/components/schemas/TagColor'}
    SegmentMembersRequest:
      type: object
      required: [recipientIds]
      properties:
        recipientIds: {type: array, minItems: 1, maxItems: 10000, items: {type: string, format: uuid}}
    RecipientSegments:
      type: object
      required: [lists, tags]
      properties:
        lists: {type: array, items: {$ref: '#/components/schemas/RecipientList'}}
        tags: {type: array, items: {$ref: '#/components/schemas/Tag'}}
...
    Problem:
      type: object
      required: [type, title, status, detail, code, category, messageKey, retryable, traceId]
      properties:
        type: {type: string, format: uri-reference}
        title: {type: string}
        status: {type: integer}
        detail: {type: string}
        code: {type: string, pattern: '^[A-Z][A-Z0-9_]*$', description: Stable application error code}
        category: {type: string, enum: [validation, authentication, permission, not_found, conflict, rate_limit, server]}
        messageKey: {type: string, description: Stable localization key}
        retryable: {type: boolean}
        fieldErrors:
          type: array
          items:
            type: object
            required: [field, code]
            properties:
              field: {type: string}
              code: {type: string}
              message: {type: string}
              value: {}
        nextAction: {type: string, description: Stable client action identifier}
        traceId: {type: string}
...
    BulkJobAction:
      type: string
      enum: [add_tag, remove_tag, add_list, remove_list, set_custom_data, export, delete]
    BulkJobCreateRequest:
      type: object
      required: [action, actionPayload, recipientIds]
      properties:
        action: {$ref: '#/components/schemas/BulkJobAction'}
        actionPayload:
          type: object
          additionalProperties: true
          description: 'set_custom_data requires key + value; tag actions require tagId; list actions require listId; export and delete require an empty object.'
        recipientIds:
          type: array
          maxItems: 100000
          items: {type: string, format: uuid}
    BulkJob:
      type: object
      required: [jobId, kind, action, status, selectionSnapshot, resolvedCount, processedRows, succeededRows, failedRows, skippedRows, idempotencyReplayed, createdAt]
      properties:
        jobId: {type: string, format: uuid}
        kind: {type: string, enum: [bulk_update]}
        action: {$ref: '#/components/schemas/BulkJobAction'}
        status: {type: string, enum: [queued, running, completed, partial_success, failed]}
        selectionSnapshot:
          type: array
          items: {type: string, format: uuid}
        resolvedCount: {type: integer, minimum: 0}
        processedRows: {type: integer, minimum: 0}
        succeededRows: {type: integer, minimum: 0}
        failedRows: {type: integer, minimum: 0}
        skippedRows: {type: integer, minimum: 0}
        idempotencyReplayed: {type: boolean}
        createdAt: {type: string, format: date-time}
    BulkJobPreview:
      type: object
      required: [scope, estimatedCount, action, actionPayload]
      properties:
        scope: {type: string, enum: [selected_recipients]}
        estimatedCount: {type: integer, minimum: 0}
        action: {$ref: '#/components/schemas/BulkJobAction'}
        actionPayload: {type: object, additionalProperties: true}
    # M5-S4. The inbound body a provider posts -- deliberately NOT named
    # ProviderWebhookEvent, which is the *parsed*, server-side shape
    # (apps/api/src/sender-config/provider-adapter.ts) that reads exactly
    # these fields and ignores every other key a payload may carry (D-106).
...
  responses:
    Problem:
      description: RFC 9457 problem
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
...